backend/setup/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1